Refactor: Split ContentView.swift into MVVM architecture
Extract 1,634-line monolithic ContentView.swift into 13 focused files:

Models:
- VideoResolution.swift: Resolution presets and validation
- FrameRate.swift: Common frame rate definitions

Services:
- FCPXMLService.swift: SRT to FCPXML conversion logic

Utilities:
- FileUtility.swift: Save dialog helper

ViewModels:
- HomeViewModel.swift: Transcription processing state
- ProcessViewModel.swift: Process view actions
- SRTConverterViewModel.swift: SRT conversion state

Views:
- TranscriptionTab.swift: Container for transcription flow
- HomeView.swift: File selection and settings UI
- ProcessView.swift: Progress and download UI
- SRTConverterView.swift: Container for SRT conversion
- SRTConverterInputView.swift: SRT file input UI
- SRTConverterResultView.swift: Conversion result UI

ContentView.swift reduced from 1,634 to 40 lines. Leverages existing
AudioService and SRTService to avoid code duplication.
