+++ Whisper Auto Captions/ContentView.swift
@@ -1,467 +1,14 @@
 import SwiftUI
-import Foundation
-import AVFoundation
-import UniformTypeIdentifiers
 #if DEBUG
 import Inject
 #endif
-
-// MARK: - Video Resolution
-enum VideoResolution: String, CaseIterable, Identifiable {
-    case hd720p = "720p HD"
-    case fullHD1080p = "1080p Full HD"
-    case uhd4K = "4K UHD"
-    case dci4K = "4K DCI"
-    case vertical1080p = "1080p Vertical"
-    case custom = "Custom"
-
-    var id: String { rawValue }
-
-    var width: Int {
-        switch self {
-        case .hd720p: return 1280
-        case .fullHD1080p: return 1920
-        case .uhd4K: return 3840
-        case .dci4K: return 4096
-        case .vertical1080p: return 1080
-        case .custom: return 1920
-        }
-    }
-
-    var height: Int {
-        switch self {
-        case .hd720p: return 720
-        case .fullHD1080p: return 1080
-        case .uhd4K, .dci4K: return 2160
-        case .vertical1080p: return 1920
-        case .custom: return 1080
-        }
-    }
-
-    var displayName: String {
-        switch self {
-        case .hd720p: return "720p HD (1280×720)"
-        case .fullHD1080p: return "1080p Full HD (1920×1080)"
-        case .uhd4K: return "4K UHD (3840×2160)"
-        case .dci4K: return "4K DCI (4096×2160)"
-        case .vertical1080p: return "1080p Vertical (1080×1920)"
-        case .custom: return "Custom"
-        }
-    }
-
-    static func formatName(width: Int, height: Int, fps: Float) -> String {
-        let fpsInt = Int(fps * 100)
-
-        switch (width, height) {
-        case (1280, 720):  return "FFVideoFormat720p\(fpsInt)"
-        case (1920, 1080): return "FFVideoFormat1080p\(fpsInt)"
-        case (3840, 2160): return "FFVideoFormat3840x2160p\(fpsInt)"
-        case (4096, 2160): return "FFVideoFormat4096x2160p\(fpsInt)"
-        case (1080, 1920): return "FFVideoFormat1080x1920p\(fpsInt)"
-        default:          return "FFVideoFormatRateUndefined"
-        }
-    }
-
-    static func isValidResolution(width: Int, height: Int) -> (valid: Bool, message: String?) {
-        if width < 640 || width > 8192 || height < 640 || height > 8192 {
-            return (false, "Resolution must be between 640 and 8192")
-        }
-        if width % 2 != 0 || height % 2 != 0 {
-            return (true, "Odd values may cause encoding issues")
-        }
-        return (true, nil)
-    }
-}
-
-// MARK: - Frame Rate
-enum FrameRate: String, CaseIterable, Identifiable {
-    case fps23_976 = "23.976"
-    case fps24 = "24"
-    case fps25 = "25"
-    case fps29_97 = "29.97"
-    case fps30 = "30"
-    case fps50 = "50"
-    case fps59_94 = "59.94"
-    case fps60 = "60"
-    case custom = "Custom"
-
-    var id: String { rawValue }
-
-    var value: Float {
-        switch self {
-        case .fps23_976: return 23.976
-        case .fps24: return 24.0
-        case .fps25: return 25.0
-        case .fps29_97: return 29.97
-        case .fps30: return 30.0
-        case .fps50: return 50.0
-        case .fps59_94: return 59.94
-        case .fps60: return 60.0
-        case .custom: return 30.0
-        }
-    }
-
-    var displayName: String {
-        switch self {
-        case .fps23_976: return "23.976 fps (Film)"
-        case .fps24: return "24 fps (Cinema)"
-        case .fps25: return "25 fps (PAL)"
-        case .fps29_97: return "29.97 fps (NTSC)"
-        case .fps30: return "30 fps"
-        case .fps50: return "50 fps (PAL HD)"
-        case .fps59_94: return "59.94 fps (NTSC HD)"
-        case .fps60: return "60 fps"
-        case .custom: return String(localized: "Custom", comment: "Custom frame rate option")
-        }
-    }
-
-    static func isValidFrameRate(_ value: Float) -> Bool {
-        return value > 0 && value <= 120
-    }
-}
-
-// MARK: - File Utility
-struct FileUtility {
-    static func saveFileWithDialog(filePath: String) {
-        let fileURL = URL(fileURLWithPath: filePath)
-        let fileManager = FileManager.default
-
-        // Verify source file exists
-        guard fileManager.fileExists(atPath: filePath) else {
-            return
-        }
-
-        let savePanel = NSSavePanel()
-        savePanel.nameFieldStringValue = fileURL.lastPathComponent
-        savePanel.canCreateDirectories = true
-        savePanel.isExtensionHidden = false
-        savePanel.title = String(localized: "Save File", comment: "Save panel title")
-        savePanel.message = String(localized: "Choose a location to save the file", comment: "Save panel message")
-
-        // Set allowed file type based on the file extension
-        let fileExtension = fileURL.pathExtension
-        if let contentType = UTType(filenameExtension: fileExtension) {
-            savePanel.allowedContentTypes = [contentType]
-        }
-
-        savePanel.begin { response in
-            if response == .OK, let destinationURL = savePanel.url {
-                do {
-                    // Remove existing file if it exists at destination
-                    if fileManager.fileExists(atPath: destinationURL.path) {
-                        try fileManager.removeItem(at: destinationURL)
-                    }
-                    try fileManager.copyItem(at: fileURL, to: destinationURL)
-                } catch {
-                    // Handle error silently
-                }
-            }
-        }
-    }
-}
-
-// MARK: - SRT Converter Utilities
-struct SRTConverter {
-    static func srtTimeToFrame(srtTime: String, fps: Float) -> Int {
-        // convert srt time to ms
-        let ms = Int(srtTime.suffix(3))!
-        let timeComponents = srtTime.prefix(srtTime.count - 4).split(separator: ":")
-        let srtTimeMs = (Int(timeComponents[0])! * 3600 + Int(timeComponents[1])! * 60 + Int(timeComponents[2])!) * 1000 + ms
-        // convert ms to frame
-        let frame = Int(floor(Float(srtTimeMs) / (1000 / fps)))
-        return frame
-    }
-
-    static func formatText(fullText: String) -> String {
-        let words = fullText.split(separator: " ")
-        var lines = [String]()
-        for i in stride(from: 0, to: words.count, by: 16) {
-            let endIndex = min(i + 16, words.count)
-            let line = words[i..<endIndex].joined(separator: " ")
-            lines.append(line)
-        }
-        let formattedText = lines.joined(separator: "\n")
-        return formattedText
-    }
-
-    static func srtToFCPXML(srtPath: String, fps: Float, projectName: String, language: String, width: Int = 1920, height: Int = 1080) -> String {
-        do {
-            let srtContent = try String(contentsOfFile: srtPath, encoding: .utf8)
-
-            // Validate SRT content is not empty
-            let trimmedContent = srtContent.trimmingCharacters(in: .whitespacesAndNewlines)
-            guard !trimmedContent.isEmpty else {
-                print("Error: SRT file is empty")
-                return "Error: SRT file is empty"
-            }
-
-            let subtitles: [String] = trimmedContent.components(separatedBy: "\n\n").filter { !$0.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty }
-
-            // Validate we have at least one subtitle
-            guard !subtitles.isEmpty else {
-                print("Error: No subtitles found in SRT file")
-                return "Error: No subtitles found"
-            }
-
-            // Validate last subtitle format before accessing
-            guard let lastSubtitle = subtitles.last else {
-                print("Error: No subtitles found in SRT file")
-                return "Error: No subtitles found"
-            }
-
-            let lastSubtitleLines = lastSubtitle.trimmingCharacters(in: .whitespacesAndNewlines).components(separatedBy: "\n")
-            guard lastSubtitleLines.count >= 2 else {
-                print("Error: Invalid subtitle format - insufficient lines")
-                return "Error: Invalid subtitle format"
-            }
-
-            let timeRangeComponents = lastSubtitleLines[1].components(separatedBy: " --> ")
-            guard timeRangeComponents.count >= 2 else {
-                print("Error: Invalid subtitle format - missing time range")
-                return "Error: Invalid subtitle format"
-            }
-
-            // extract total duration from srt
-            let totalSrtTime = timeRangeComponents[1]
-            let totalFrame = srtTimeToFrame(srtTime: totalSrtTime, fps: Float(fps))
-            let hundredFoldTotalFrame = String(100 * totalFrame)
-            let hundredFoldFps = String(Int(fps * 100))
-
-            // fcpxml
-            let fcpxmlElement = XMLElement(name: "fcpxml")
-            fcpxmlElement.addAttribute(XMLNode.attribute(withName: "version", stringValue: "1.9") as! XMLNode)
-
-            // resource
-            let resourcesElement = XMLElement(name: "resources")
-
-            // format - use dynamic resolution
-            let formatName = VideoResolution.formatName(width: width, height: height, fps: fps)
-            let formatElement = XMLElement(name: "format")
-            formatElement.addAttribute(XMLNode.attribute(withName: "id", stringValue: "r1") as! XMLNode)
-            formatElement.addAttribute(XMLNode.attribute(withName: "name", stringValue: formatName) as! XMLNode)
-            formatElement.addAttribute(XMLNode.attribute(withName: "frameDuration", stringValue: "100/\(hundredFoldFps)s") as! XMLNode)
-            formatElement.addAttribute(XMLNode.attribute(withName: "width", stringValue: String(width)) as! XMLNode)
-            formatElement.addAttribute(XMLNode.attribute(withName: "height", stringValue: String(height)) as! XMLNode)
-            formatElement.addAttribute(XMLNode.attribute(withName: "colorSpace", stringValue: "1-1-1 (Rec. 709)") as! XMLNode)
-            resourcesElement.addChild(formatElement)
-
-            // effect
-            let effectElement = XMLElement(name: "effect")
-            effectElement.addAttribute(XMLNode.attribute(withName: "id", stringValue: "r2") as! XMLNode)
-            effectElement.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Basic Title") as! XMLNode)
-            effectElement.addAttribute(XMLNode.attribute(withName: "uid", stringValue: ".../Titles.localized/Bumper:Opener.localized/Basic Title.localized/Basic Title.moti") as! XMLNode)
-            resourcesElement.addChild(effectElement)
-
-            // library
-            let libraryElement = XMLElement(name: "library")
-
-            // event
-            let eventElement = XMLElement(name: "event")
-            eventElement.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Whisper Auto Captions") as! XMLNode)
-            libraryElement.addChild(eventElement)
-
-            // project
-            let projectElement = XMLElement(name: "project")
-            projectElement.addAttribute(XMLNode.attribute(withName: "name", stringValue: "\(projectName)") as! XMLNode)
-            eventElement.addChild(projectElement)
-
-            // sequence
-            let sequenceElement = XMLElement(name: "sequence")
-            sequenceElement.addAttribute(XMLNode.attribute(withName: "format", stringValue: "r1") as! XMLNode)
-            sequenceElement.addAttribute(XMLNode.attribute(withName: "tcStart", stringValue: "0s") as! XMLNode)
-            sequenceElement.addAttribute(XMLNode.attribute(withName: "tcFormat", stringValue: "NDF") as! XMLNode)
-            sequenceElement.addAttribute(XMLNode.attribute(withName: "audioLayout", stringValue: "stereo") as! XMLNode)
-            sequenceElement.addAttribute(XMLNode.attribute(withName: "audioRate", stringValue: "48k") as! XMLNode)
-            sequenceElement.addAttribute(XMLNode.attribute(withName: "duration", stringValue: "\(totalFrame)/\(hundredFoldFps)s") as! XMLNode)
-            projectElement.addChild(sequenceElement)
-
-            // spine
-            let spineElement = XMLElement(name: "spine")
-            sequenceElement.addChild(spineElement)
-
-            // gap
-            let gapElement = XMLElement(name: "gap")
-            gapElement.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Gap") as! XMLNode)
-            gapElement.addAttribute(XMLNode.attribute(withName: "offset", stringValue: "0s") as! XMLNode)
-            gapElement.addAttribute(XMLNode.attribute(withName: "duration", stringValue: "\(hundredFoldTotalFrame)/\(hundredFoldFps)s") as! XMLNode)
-            spineElement.addChild(gapElement)
-
-            for (i, subtitle) in subtitles.enumerated() {
-                let subtitleItem = subtitle.trimmingCharacters(in: .whitespacesAndNewlines).components(separatedBy: "\n")
-
-                // Skip malformed subtitles (need at least: index, time range, text)
-                guard subtitleItem.count >= 3 else {
-                    print("Warning: Skipping malformed subtitle at index \(i)")
-                    continue
-                }
-
-                let timeRange = subtitleItem[1].components(separatedBy: " --> ")
-                guard timeRange.count >= 2 else {
-                    print("Warning: Skipping subtitle with invalid time range at index \(i)")
-                    continue
-                }
-
-                let offset = timeRange[0]
-                let end = timeRange[1]
-                let offsetFrame = srtTimeToFrame(srtTime: offset, fps: fps)
-                let endFrame = srtTimeToFrame(srtTime: end, fps: fps)
-                let durationFrame = endFrame - offsetFrame
-
-                let hundredFoldOffsetFrame = String(100 * offsetFrame)
-                let hundredFoldDurationFrame = String(100 * durationFrame)
-                var subtitleContent = subtitleItem[2]
-                if language == "English" {
-                    if subtitleContent.split(separator: " ").count > 16 {
-                        subtitleContent = formatText(fullText: subtitleContent)
-                    }
-                }
-
-                if language == "Chinese" {
-                    // title
-                    let titleElement = XMLElement(name: "title")
-                    titleElement.addAttribute(XMLNode.attribute(withName: "ref", stringValue: "r2") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "lane", stringValue: "1") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "offset", stringValue: "\(hundredFoldOffsetFrame)/\(hundredFoldFps)s") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "duration", stringValue: "\(hundredFoldDurationFrame)/\(hundredFoldFps)s") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "name", stringValue: "\(subtitleContent) - Basic Title") as! XMLNode)
-
-                    // param1
-                    let param1Element = XMLElement(name: "param")
-                    param1Element.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Position") as! XMLNode)
-                    param1Element.addAttribute(XMLNode.attribute(withName: "key", stringValue: "9999/999166631/999166633/1/100/101") as! XMLNode)
-                    param1Element.addAttribute(XMLNode.attribute(withName: "value", stringValue: "0 -465") as! XMLNode)
-                    titleElement.addChild(param1Element)
-
-                    // param2
-                    let param2Element = XMLElement(name: "param")
-                    param2Element.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Flatten") as! XMLNode)
-                    param2Element.addAttribute(XMLNode.attribute(withName: "key", stringValue: "999/999166631/999166633/2/351") as! XMLNode)
-                    param2Element.addAttribute(XMLNode.attribute(withName: "value", stringValue: "1") as! XMLNode)
-                    titleElement.addChild(param2Element)
-
-                    // param3
-                    let param3Element = XMLElement(name: "param")
-                    param3Element.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Alignment") as! XMLNode)
-                    param3Element.addAttribute(XMLNode.attribute(withName: "key", stringValue: "9999/999166631/999166633/2/354/999169573/401") as! XMLNode)
-                    param3Element.addAttribute(XMLNode.attribute(withName: "value", stringValue: "1 (Center)") as! XMLNode)
-                    titleElement.addChild(param3Element)
-
-                    // text
-                    let textElement = XMLElement(name: "text")
-                    titleElement.addChild(textElement)
-
-                    // text style
-                    let textStyleElement = XMLElement(name: "text-style")
-                    textStyleElement.addAttribute(XMLNode.attribute(withName: "ref", stringValue: "ts\(String(i))") as! XMLNode)
-                    textStyleElement.stringValue = subtitleContent
-                    textElement.addChild(textStyleElement)
-
-                    // text style def
-                    let textStyleDefElement = XMLElement(name: "text-style-def")
-                    textStyleDefElement.addAttribute(XMLNode.attribute(withName: "id", stringValue: "ts\(String(i))") as! XMLNode)
-                    titleElement.addChild(textStyleDefElement)
-
-                    // text style 2
-                    let textStyle2Element = XMLElement(name: "text-style")
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "font", stringValue: "PingFang SC") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "fontSize", stringValue: "50") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "fontFace", stringValue: "Semibold") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "fontColor", stringValue: "1 1 1 1") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "bold", stringValue: "1") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "shadowColor", stringValue: "0 0 0 0.75") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "shadowOffset", stringValue: "4 315") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "alignment", stringValue: "center") as! XMLNode)
-                    textStyleDefElement.addChild(textStyle2Element)
-
-                    gapElement.addChild(titleElement)
-
-                } else {
-                    // title
-                    let titleElement = XMLElement(name: "title")
-                    titleElement.addAttribute(XMLNode.attribute(withName: "ref", stringValue: "r2") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "lane", stringValue: "1") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "offset", stringValue: "\(hundredFoldOffsetFrame)/\(hundredFoldFps)s") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "duration", stringValue: "\(hundredFoldDurationFrame)/\(hundredFoldFps)s") as! XMLNode)
-                    titleElement.addAttribute(XMLNode.attribute(withName: "name", stringValue: "\(subtitleContent) - Basic Title") as! XMLNode)
-
-                    // param1
-                    let param1Element = XMLElement(name: "param")
-                    param1Element.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Position") as! XMLNode)
-                    param1Element.addAttribute(XMLNode.attribute(withName: "key", stringValue: "9999/999166631/999166633/1/100/101") as! XMLNode)
-                    param1Element.addAttribute(XMLNode.attribute(withName: "value", stringValue: "0 -465") as! XMLNode)
-                    titleElement.addChild(param1Element)
-
-                    // param2
-                    let param2Element = XMLElement(name: "param")
-                    param2Element.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Flatten") as! XMLNode)
-                    param2Element.addAttribute(XMLNode.attribute(withName: "key", stringValue: "999/999166631/999166633/2/351") as! XMLNode)
-                    param2Element.addAttribute(XMLNode.attribute(withName: "value", stringValue: "1") as! XMLNode)
-                    titleElement.addChild(param2Element)
-
-                    // param3
-                    let param3Element = XMLElement(name: "param")
-                    param3Element.addAttribute(XMLNode.attribute(withName: "name", stringValue: "Alignment") as! XMLNode)
-                    param3Element.addAttribute(XMLNode.attribute(withName: "key", stringValue: "9999/999166631/999166633/2/354/999169573/401") as! XMLNode)
-                    param3Element.addAttribute(XMLNode.attribute(withName: "value", stringValue: "1 (Center)") as! XMLNode)
-                    titleElement.addChild(param3Element)
-
-                    // text
-                    let textElement = XMLElement(name: "text")
-                    titleElement.addChild(textElement)
-
-                    // text style
-                    let textStyleElement = XMLElement(name: "text-style")
-                    textStyleElement.addAttribute(XMLNode.attribute(withName: "ref", stringValue: "ts\(String(i))") as! XMLNode)
-                    textStyleElement.stringValue = subtitleContent
-                    textElement.addChild(textStyleElement)
-
-                    // text style def
-                    let textStyleDefElement = XMLElement(name: "text-style-def")
-                    textStyleDefElement.addAttribute(XMLNode.attribute(withName: "id", stringValue: "ts\(String(i))") as! XMLNode)
-                    titleElement.addChild(textStyleDefElement)
-
-                    // text style 2
-                    let textStyle2Element = XMLElement(name: "text-style")
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "font", stringValue: "Helvetica") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "fontSize", stringValue: "45") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "fontFace", stringValue: "Regular") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "fontColor", stringValue: "1 1 1 1") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "shadowColor", stringValue: "0 0 0 0.75") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "shadowOffset", stringValue: "4 315") as! XMLNode)
-                    textStyle2Element.addAttribute(XMLNode.attribute(withName: "alignment", stringValue: "center") as! XMLNode)
-                    textStyleDefElement.addChild(textStyle2Element)
-
-                    gapElement.addChild(titleElement)
-                }
-            }
-
-            // Add the resources and library elements to the fcpxml element
-            fcpxmlElement.addChild(resourcesElement)
-            fcpxmlElement.addChild(libraryElement)
-
-            // Create the XML document with the fcpxml element as the root
-            let xmlDoc = XMLDocument(rootElement: fcpxmlElement)
-
-            // Set the XML document version and encoding
-            xmlDoc.version = "1.0"
-            xmlDoc.characterEncoding = "utf-8"
-
-            // Write the XML document to the output file
-            let xmlData = xmlDoc.xmlData(options: .nodePrettyPrint)
-            let fileUrl = URL(fileURLWithPath: srtPath + ".fcpxml")
-            try! xmlData.write(to: fileUrl)
-            return srtPath + ".fcpxml"
-        } catch {
-            return "Error"
-        }
-    }
-}
 
 // MARK: - Main Content View with Tabs
 struct ContentView: View {
     #if DEBUG
     @ObserveInjection var inject
     #endif
+    
     @State private var selectedTab = 0
 
     var body: some View {
@@ -479,1156 +26,15 @@ struct ContentView: View {
                 .tag(1)
         }
         .frame(minWidth: 600, minHeight: 600)
-
+        #if DEBUG
+        .enableInjection()
+        #endif
     }
 }
 
-// MARK: - Transcription Tab (Original Functionality)
-struct TranscriptionTab: View {
-    #if DEBUG
-    @ObserveInjection var inject
-    #endif
-    @State var startCreatingAutoCaptions = false
-    @State var progress = 0.0
-    @State var progressPercentage = 0
-    @State var totalBatch = 100000
-    @State var currentBatch = -100000
-    @State var remainingTime = "00:00"
-    @State var status = "Spliting audio file by every 10 minutes···"
-    @State var projectName = ""
-    @State var outputCaptions = ""
-    @State var outputSRTFilePath = ""
-    @State var outputFCPXMLFilePath = ""
-
-    var body: some View {
-        Group {
-            if startCreatingAutoCaptions {
-                ProcessView(progress: $progress, progressPercentage: $progressPercentage, totalBatch: $totalBatch, currentBatch: $currentBatch, remainingTime: $remainingTime, status: $status, outputCaptions: $outputCaptions, projectName: $projectName, outputFCPXMLFilePath: $outputFCPXMLFilePath, outputSRTFilePath: $outputSRTFilePath, isPresented: $startCreatingAutoCaptions)
-            } else {
-                HomeView(startCreatingAutoCaptions: $startCreatingAutoCaptions, progress: $progress, progressPercentage: $progressPercentage, totalBatch: $totalBatch, currentBatch: $currentBatch, remainingTime: $remainingTime, status: $status, outputCaptions: $outputCaptions, projectName: $projectName, outputFCPXMLFilePath: $outputFCPXMLFilePath, outputSRTFilePath: $outputSRTFilePath)
-            }
-        }
-
-    }
-}
-
-
-
-struct HomeView: View {
-    #if DEBUG
-    @ObserveInjection var inject
-    #endif
-    @State var fileURL: URL?
-    @State var isSelected: Bool = false
-    @State private var selectedFrameRate: FrameRate = .fps30
-    @State private var customFps: String = "30"
-    @State private var selectedLanguage = "Auto"
-    @State private var selectedModel = "Medium"
-    @State private var selectedPreset: WhisperPreset = .balanced
-    @State private var showSettings = false
-
-    private var currentFps: Float {
-        if selectedFrameRate == .custom {
-            return Float(customFps) ?? 30.0
-        }
-        return selectedFrameRate.value
-    }
-
-    private var isFpsValid: Bool {
-        return FrameRate.isValidFrameRate(currentFps)
-    }
-
-    let languages = LanguageData.languages
-    let languagesMapping = LanguageData.languageToCode
-    let models = ModelData.models
-    let modelsMapping = ModelData.modelToFileName
-
-    @ObservedObject var settingsManager = SettingsManager.shared
-
-    @State var fileName: String = ""
-
-    @Binding var startCreatingAutoCaptions: Bool
-    @Binding var progress: Double
-    @Binding var progressPercentage: Int
-    @Binding var totalBatch: Int
-    @Binding var currentBatch: Int
-    @Binding var remainingTime: String
-    @Binding var status: String
-    @Binding var outputCaptions: String
-    @Binding var projectName: String
-    @Binding var outputFCPXMLFilePath: String
-    @Binding var outputSRTFilePath: String
-
-    @State private var isDownloading = false
-    @State private var downloadProgress: Double = 0.0
-    @State private var showAlert = false
-    @State private var downloadDelegate: DownloadDelegate?
-    @State private var downloadTask: URLSessionDownloadTask?
-
-    
-    var body: some View {
-        VStack {
-            Grid(alignment: .leadingFirstTextBaseline, verticalSpacing: 20) {
-                GridRow {
-                        Text("Audio File:")
-                            .gridColumnAlignment(.trailing)
-                        HStack {
-                            Button(action: {
-                                let panel = NSOpenPanel()
-                                panel.canChooseFiles = true
-                                panel.canChooseDirectories = true
-                                panel.allowsMultipleSelection = false
-                                panel.allowedContentTypes = [.audio, .movie]
-                                if panel.runModal() == .OK {
-                                    self.fileURL = panel.urls.first
-                                    if let fileName = panel.urls.first?.lastPathComponent {
-                                        self.fileName = fileName
-                                        self.projectName = (fileName as NSString).deletingPathExtension
-                                    }
-                                }
-                            }) {
-                                Text("Choose File")
-                            }
-
-                            if self.fileURL != nil {
-                                Text(fileName)
-                            }
-                        }
-                        .gridColumnAlignment(.leading)
-                }
-
-                GridRow {
-                    Text(String(localized: "Frame Rate:", comment: "Frame rate label"))
-                    VStack(alignment: .leading, spacing: 8) {
-                        Picker(selection: $selectedFrameRate, label: EmptyView()) {
-                            ForEach(FrameRate.allCases) { frameRate in
-                                Text(frameRate.displayName).tag(frameRate)
-                            }
-                        }
-                        .pickerStyle(MenuPickerStyle())
-                        .frame(alignment:.leading)
-
-                        if selectedFrameRate == .custom {
-                            HStack(spacing: 8) {
-                                TextField("fps", text: $customFps)
-                                    .textFieldStyle(RoundedBorderTextFieldStyle())
-                                    .frame(width: 70)
-                                Text("fps")
-                            }
-
-                            if !isFpsValid {
-                                Text(String(localized: "Frame rate must be between 0 and 120", comment: "Invalid frame rate warning"))
-                                    .font(.caption)
-                                    .foregroundColor(.red)
-                            }
-                        }
-                    }
-                }
-
-                GridRow {
-                    Text(String(localized: "Model:", comment: "Model label"))
-                    Picker(selection: $selectedModel, label: EmptyView()) {
-                        ForEach(models, id: \.self) {
-                            Text($0)
-                        }
-                    }
-                    .pickerStyle(MenuPickerStyle())
-                    .frame(alignment: .leading)
-                }
-
-                GridRow {
-                    Text(String(localized: "Language:", comment: "Language label"))
-                    Picker(selection: $selectedLanguage, label: EmptyView()) {
-                        ForEach(languages, id: \.self) {
-                            Text($0)
-                        }
-                    }
-                    .pickerStyle(MenuPickerStyle())
-                    .frame(alignment: .leading)
-                }
-
-                GridRow {
-                    Text(String(localized: "Preset:", comment: "Preset label"))
-                    Picker(selection: $selectedPreset, label: EmptyView()) {
-                        ForEach(WhisperPreset.allCases) { preset in
-                            Text(preset.displayName).tag(preset)
-                        }
-                    }
-                    .pickerStyle(MenuPickerStyle())
-                    .frame(alignment: .leading)
-                    .onChange(of: selectedPreset) { newPreset in
-                        settingsManager.currentPreset = newPreset
-                    }
-                }
-
-                GridRow {
-                    HStack(spacing: 12) {
-                        Button(action: { showSettings = true }) {
-                            HStack {
-                                Image(systemName: "gear")
-                                Text(String(localized: "Settings", comment: "Settings button"))
-                            }
-                        }
-
-                        Button(action: {
-                        let fileManager = FileManager.default
-                        let applicationSupportDirectory = try! fileManager.url(for: .applicationSupportDirectory, in: .userDomainMask, appropriateFor: nil, create: true)
-                        let whisperAutoCaptionsURL = applicationSupportDirectory.appendingPathComponent("Whisper Auto Captions")
-                        try? fileManager.createDirectory(at: whisperAutoCaptionsURL, withIntermediateDirectories: true, attributes: nil)
-                        let destinationURL = whisperAutoCaptionsURL.appendingPathComponent("ggml-\(modelsMapping[selectedModel]!).bin")
-
-                        if fileManager.fileExists(atPath: destinationURL.path) {
-                            // Validate existing model file
-                            do {
-                                let fileAttributes = try fileManager.attributesOfItem(atPath: destinationURL.path)
-                                let fileSize = fileAttributes[.size] as? Int64 ?? 0
-                                let minimumValidFileSize: Int64 = 50 * 1024 * 1024  // 50MB minimum
-                                
-                                if fileSize >= minimumValidFileSize {
-                                    // Model file is valid, proceed
-                                    whisper_auto_captions()
-                                } else {
-                                    // Model file is invalid/corrupted, delete and re-download
-                                    print("Invalid model file detected (\(fileSize) bytes), deleting and re-downloading...")
-                                    try fileManager.removeItem(atPath: destinationURL.path)
-                                    download_model(model: modelsMapping[selectedModel]!) { success in
-                                        if success {
-                                            whisper_auto_captions()
-                                        }
-                                    }
-                                }
-                            } catch {
-                                // Can't read file attributes, treat as invalid
-                                print("Failed to validate model file: \(error), re-downloading...")
-                                try? fileManager.removeItem(atPath: destinationURL.path)
-                                download_model(model: modelsMapping[selectedModel]!) { success in
-                                    if success {
-                                        whisper_auto_captions()
-                                    }
-                                }
-                            }
-                        } else {
-//                            print("File does not exist")
-                            download_model(model: modelsMapping[selectedModel]!) { success in
-                                if success {
-                                    whisper_auto_captions()
-                                } else {
-
-                                }
-                            }
-                        }
-                        }, label: {
-                            Text(String(localized: "Create", comment: "Create button"))
-                        }).buttonStyle(BorderedProminentButtonStyle())
-                            .disabled(fileURL == nil || !isFpsValid)
-                    }
-                }.gridCellColumns(2)
-                    .gridCellAnchor(.center)
-            }
-            if isDownloading {
-                ProgressView(value: downloadProgress)
-                    .padding()
-                    .progressViewStyle(LinearProgressViewStyle())
-                    .frame(width: 200)
-            }
-        }.alert(isPresented: $showAlert) {
-            Alert(
-                title: Text("Downloading \(selectedModel) Model"),
-                message: Text(String(format: "Progress: %.0f%%", downloadProgress * 100)),
-                primaryButton: .destructive(Text("Cancel"), action: {
-                    // Cancel the download task
-                    self.downloadDelegate?.cancelAction?()
-                    self.isDownloading = false
-                    self.showAlert = false
-                }),
-                secondaryButton: .default(Text(""), action: {})
-            )
-        }
-        .sheet(isPresented: $showSettings) {
-            SettingsWindowView()
-        }
-
-    }
-
-    func whisper_auto_captions() {
-        self.startCreatingAutoCaptions = true
-        let filePathString = fileURL!.path
-        let tempFolder = NSTemporaryDirectory()
-
-        // convert mp3 to 16kHz wav file
-        let outputWavFilePath = mp3_to_wav(filePathString: filePathString, projectName: projectName, tempFolder: tempFolder)
-
-        let splitedWavFilesPaths = spilt_wav(inputFilePath: outputWavFilePath)
-        self.totalBatch = splitedWavFilesPaths.count
-        self.status = "Generating AI subtitles"
-
-        var srtFiles = [String]()
-
-        DispatchQueue.global(qos: .background).async {
-            let group = DispatchGroup()
-
-            for (b, splitedWavFilePath) in splitedWavFilesPaths.enumerated() {
-                self.currentBatch = b + 1
-                var outputSplitSRTFilePath: String?
-                group.enter()
-                whisper_cpp(selectedModel: selectedModel, selectedLanguage: selectedLanguage, outputWavFilePath: splitedWavFilePath) { srtFilePath in
-                    outputSplitSRTFilePath = srtFilePath
-                    group.leave()
-                }
-
-                group.wait()
-
-                if let srtFilePath = outputSplitSRTFilePath, !srtFilePath.isEmpty {
-                    DispatchQueue.main.async {
-                        srtFiles.append(srtFilePath)
-                        self.progress = 0.0
-                        self.progressPercentage = 0
-                    }
-                } else {
-                    print("Warning: Batch \(b + 1) failed to generate SRT file")
-                }
-            }
-
-            DispatchQueue.main.async {
-                self.progress = 1.0
-                self.progressPercentage = 100
-
-                if srtFiles.isEmpty {
-                    self.status = "Error: No subtitles generated"
-                    self.startCreatingAutoCaptions = false
-                    return
-                }
-
-                let outputSRTFilePath = merge_srt(srt_files: srtFiles)
-                // srt to fcpxml
-                self.status = "Done"
-                self.outputFCPXMLFilePath = SRTConverter.srtToFCPXML(srtPath: outputSRTFilePath, fps: self.currentFps, projectName: projectName, language: selectedLanguage)
-
-                self.outputSRTFilePath = outputSRTFilePath
-            }
-        }
-    }
-
-
-    func spilt_wav(inputFilePath: String) -> [String] {
-        var result: [String] = []
-        // Append file:// to the beginning
-        let fileURL = URL(fileURLWithPath: inputFilePath)
-        if let asset: AVAudioPlayer = try? AVAudioPlayer(contentsOf: fileURL), asset.duration > 0 {
-            let duration: TimeInterval = asset.duration
-            let segmentDuration: Double = 600
-            if duration < segmentDuration {
-                return [inputFilePath]
-            }
-            let numberOfSegments = Int(ceil(duration / segmentDuration) - 1)
-            
-            let outputPrefix = "\(inputFilePath)_p"
-            for i in 0...numberOfSegments {
-                let outputFilePath = "\(outputPrefix)\(i).wav"
-                
-                let task = Process()
-                let ffmpegPath = Bundle.main.path(forResource: "ffmpeg", ofType: nil)
-                task.launchPath = ffmpegPath
-                task.arguments = ["-i", inputFilePath, "-ss", String(i*Int(segmentDuration)), "-t", String(Int(segmentDuration)), "-c", "copy", outputFilePath]
-                task.launch()
-                task.waitUntilExit()
-                result.append(outputFilePath)
-            }
-        } else {
-            // Fallback: return original file if can't determine duration
-            return [inputFilePath]
-        }
-        return result
-    }
-    
-    func adjustSrtTime(srtTime: String, factor: Int) -> String {
-        let timeComponents = srtTime.components(separatedBy: ":")
-        let hours = Int(timeComponents[0]) ?? 0
-        let minutes = Int(timeComponents[1]) ?? 0
-        let secondsAndMilliseconds = timeComponents[2].components(separatedBy: ",")
-        let seconds = Int(secondsAndMilliseconds[0]) ?? 0
-        
-        let totalMinutes = (hours * 60) + minutes
-        let newTotalMinutes = totalMinutes + (factor * 10)
-        let newHours = newTotalMinutes / 60
-        let newMinutes = newTotalMinutes % 60
-        
-        let newTime = String(format: "%02d:%02d:%02d,%@", newHours, newMinutes, seconds, secondsAndMilliseconds[1])
-        
-        return newTime
-    }
-
-
-    func merge_srt(srt_files: [String]) -> String {
-        guard !srt_files.isEmpty else {
-            return ""
-        }
-        let merged_srt_path = srt_files[0] + "_merged.srt"
-        var merged_contents = ""
-        var index = 1
-        for (i, srt_path) in srt_files.enumerated() {
-            do {
-                let srt_content = try String(contentsOfFile: srt_path, encoding: .utf8)
-                let trimmed_content = srt_content.trimmingCharacters(in: .whitespacesAndNewlines)
-
-                // Skip empty SRT files
-                guard !trimmed_content.isEmpty else {
-                    print("Warning: Skipping empty SRT file: \(srt_path)")
-                    continue
-                }
-
-                let subtitles: [String] = trimmed_content.components(separatedBy: "\n\n").filter { !$0.trimmingCharacters(in: .whitespacesAndNewlines).isEmpty }
-                for (_, subtitle) in subtitles.enumerated() {
-                    let subtitle_item = subtitle.trimmingCharacters(in: .whitespacesAndNewlines).components(separatedBy: "\n")
-
-                    // Skip malformed subtitles
-                    guard subtitle_item.count >= 3 else {
-                        print("Warning: Skipping malformed subtitle in \(srt_path)")
-                        continue
-                    }
-
-                    let time_range = subtitle_item[1].components(separatedBy: " --> ")
-                    guard time_range.count >= 2 else {
-                        print("Warning: Skipping subtitle with invalid time range in \(srt_path)")
-                        continue
-                    }
-
-                    let start = time_range[0]
-                    let end = time_range[1]
-
-                    let new_start = adjustSrtTime(srtTime: start, factor: i)
-                    let new_end = adjustSrtTime(srtTime: end, factor: i)
-
-                    let new_time_range = new_start + " --> " + new_end
-                    let subtitle_content = subtitle_item[2]
-
-                    merged_contents += "\(index)\n"
-                    merged_contents += "\(new_time_range)\n"
-                    merged_contents += "\(subtitle_content)\n\n"
-                    index += 1
-
-                }
-            } catch {
-                print("Error reading SRT file \(srt_path): \(error)")
-            }
-        }
-        do {
-            try merged_contents.write(toFile: merged_srt_path, atomically: true, encoding: .utf8)
-
-        } catch {
-//            print("write with error：\(error)")
-        }
-        return merged_srt_path
-    }
-    
-    func download_model(model: String, completion: @escaping (Bool) -> Void) {
-        guard let url = URL(string: "https://huggingface.co/ggerganov/whisper.cpp/resolve/main/ggml-\(model.lowercased()).bin") else {
-            completion(false)
-            return
-        }
-
-        self.downloadDelegate = DownloadDelegate(model: selectedModel, progressHandler: { progress in
-            self.downloadProgress = progress
-        }, completionHandler: {
-            self.isDownloading = false
-            self.showAlert = false
-            completion(true)
-        }, errorHandler: { errorMessage in
-            self.isDownloading = false
-            self.showAlert = false
-            print("Download error: \(errorMessage)")
-            completion(false)
-        })
-        
-
-        self.downloadDelegate?.cancelAction = {
-            self.downloadProgress = 0.0
-            self.downloadTask?.cancel()
-        }
-    
-        let session = URLSession(configuration: .default, delegate: self.downloadDelegate, delegateQueue: nil)
-        let task = session.downloadTask(with: url)
-        self.downloadTask = task
-
-        self.isDownloading = true
-        self.showAlert = true
-
-        task.resume()
-    }
-    
- 
-    func mp3_to_wav(filePathString: String, projectName: String, tempFolder: String) -> String {
-        var wavFileName = projectName + ".wav"
-        var wavFilePath = tempFolder + wavFileName
-
-        let fileManager = FileManager.default
-        if fileManager.fileExists(atPath: wavFilePath) {
-            // Rename the file if it already exists
-            var counter = 1
-            while fileManager.fileExists(atPath: wavFilePath) {
-                wavFileName = "\(projectName)_\(counter).wav"
-                wavFilePath = tempFolder + wavFileName
-                counter += 1
-            }
-        }
-
-        let ffmpegPath = Bundle.main.path(forResource: "ffmpeg", ofType: nil)
-        let task = Process()
-        task.launchPath = ffmpegPath
-        task.arguments = ["-i", filePathString, "-ar", "16000", "-ac", "1", wavFilePath]
-        task.launch()
-        task.waitUntilExit()
-        // Note: task.terminationStatus can be checked here if error handling is needed
-        return wavFilePath
-    }
-    
-    func whisper_cpp(selectedModel: String, selectedLanguage: String, outputWavFilePath: String, completion: @escaping (String) -> Void) {
-        DispatchQueue.global(qos: .background).async {
-            let fileManager = FileManager.default
-            let applicationSupportDirectory = try! fileManager.url(for: .applicationSupportDirectory, in: .userDomainMask, appropriateFor: nil, create: true)
-            let whisperAutoCaptionsURL = applicationSupportDirectory.appendingPathComponent("Whisper Auto Captions")
-            let modelPath = whisperAutoCaptionsURL.appendingPathComponent("ggml-\(selectedModel.lowercased()).bin")
-            
-            
-            if let whisperCliPath = Bundle.main.path(forResource: "whisper-cli", ofType: nil),
-               let selectedLanguageShortCut = languagesMapping[selectedLanguage]
-            {
-                let task = Process()
-                task.launchPath = whisperCliPath
-                if selectedLanguageShortCut == "auto" {
-                    // Auto mode: let whisper.cpp detect language automatically (no prompt)
-                    task.arguments = ["-m", modelPath.path, "-l", "auto", "-pp", "-osrt", "-f", outputWavFilePath]
-                } else if selectedLanguageShortCut != "zh" {
-                    task.arguments = ["-m", modelPath.path, "-l", selectedLanguageShortCut, "-pp", "-osrt", "-f", outputWavFilePath]
-                } else {
-                    if selectedLanguage == "Chinese Simplified" {
-                        task.arguments = ["-m", modelPath.path, "-l", selectedLanguageShortCut, "-pp", "-osrt", "-f", outputWavFilePath, "--prompt", "\"以下是普通话的句子\""]
-                    } else {
-                        task.arguments = ["-m", modelPath.path, "-l", selectedLanguageShortCut, "-pp", "-osrt", "-f", outputWavFilePath, "--prompt", "\"以下是普通話的句子\""]
-                    }
-                }
-
-            
-                let errorPipe = Pipe()
-                let outputPipe = Pipe()
-                
-                task.standardError = errorPipe
-                task.standardOutput = outputPipe
-                
-                task.launch()
-                let startTime = Date()
-                
-                let errorHandle = errorPipe.fileHandleForReading
-                let outputHandle = outputPipe.fileHandleForReading
-                
-                
-                while task.isRunning || errorHandle.availableData.count > 0 {
-                    let errorData = errorHandle.availableData
-                    if !errorData.isEmpty {
-                        if let error = String(data: errorData, encoding: .utf8) {
-                            let lines = error.split(separator: "\n")
-                            if let lastLine = lines.last, lastLine.hasPrefix("whisper_full_with_state: progress") {
-                                if let progressString = lastLine.components(separatedBy: "=").last?.trimmingCharacters(in: .whitespacesAndNewlines).dropLast() {
-                                    let progressPercentage = Int(progressString) ?? 0
-                                    let progress = Double(Int(progressString) ?? 0) * 0.01
-                                    DispatchQueue.main.async {
-                                        self.progressPercentage = progressPercentage
-                                        self.progress = progress
-                                    }
-                                    let currentTime = Date()
-                                    let elapsed = currentTime.timeIntervalSince(startTime)
-                                    let remainingSeconds = round((1 - Double(progress)) / Double(progress) * elapsed)
-                                    let remainingTime = formatSeconds(remainingSeconds)
-                                    DispatchQueue.main.async {
-                                        self.remainingTime = remainingTime
-                                    }
-//                                    print("my current progerss is", progress, remainingTime)
-                                }
-                            }
-                        }
-                    }
-                    let outputData = outputHandle.availableData
-                    if !outputData.isEmpty {
-                        if let outputCaptions = String(data: outputData, encoding: .utf8) {
-                            DispatchQueue.main.async {
-                                self.progress = progress
-                                self.progressPercentage = progressPercentage
-                                self.remainingTime = remainingTime
-                                self.outputCaptions += outputCaptions
-                            }
-                        }
-                    }
-                }
-                
-                
-                task.waitUntilExit()
-
-                DispatchQueue.main.async {
-                    self.progress = 1.0
-                    self.progressPercentage = 100
-                    self.remainingTime = "00:00"
-                }
-
-                let outputData = outputHandle.readDataToEndOfFile()
-                _ = String(data: outputData, encoding: .utf8)
-
-
-            }
-
-            let srtFilePath = outputWavFilePath + ".srt"
-
-            // Verify the SRT file was created and has content (reusing fileManager from above)
-            if fileManager.fileExists(atPath: srtFilePath) {
-                do {
-                    let fileAttributes = try fileManager.attributesOfItem(atPath: srtFilePath)
-                    let fileSize = fileAttributes[.size] as? Int64 ?? 0
-
-                    if fileSize > 0 {
-                        // Check if the file has valid SRT content (at least one subtitle entry)
-                        let srtContent = try String(contentsOfFile: srtFilePath, encoding: .utf8)
-                        let trimmed = srtContent.trimmingCharacters(in: .whitespacesAndNewlines)
-                        if !trimmed.isEmpty && trimmed.contains("-->") {
-                            completion(srtFilePath)
-                        } else {
-                            print("Error: SRT file generated but contains no valid subtitles")
-                            completion("")
-                        }
-                    } else {
-                        print("Error: SRT file is empty - whisper-cli may have failed")
-                        completion("")
-                    }
-                } catch {
-                    print("Error reading SRT file: \(error)")
-                    completion("")
-                }
-            } else {
-                print("Error: SRT file was not created - whisper-cli may have failed")
-                completion("")
-            }
-        }
-        }
-
-    func formatSeconds(_ seconds: Double) -> String {
-        let formatter = DateComponentsFormatter()
-        formatter.allowedUnits = [.hour, .minute, .second]
-        formatter.unitsStyle = .positional
-        formatter.zeroFormattingBehavior = .pad
-        if seconds >= 3600 {
-            formatter.maximumUnitCount = 3
-        } else  {
-            formatter.maximumUnitCount = 2
-            formatter.allowedUnits = [.minute, .second]
-        }
-        let formattedTime = formatter.string(from: seconds)!
-        return formattedTime
-    }
-}
-
-struct ProcessView: View {
-    #if DEBUG
-    @ObserveInjection var inject
-    #endif
-    @Binding var progress: Double
-    @Binding var progressPercentage: Int
-    @Binding var totalBatch: Int
-    @Binding var currentBatch: Int
-    @Binding var remainingTime: String
-    @Binding var status: String
-    @Binding var outputCaptions: String
-    @Binding var projectName: String
-    @Binding var outputFCPXMLFilePath: String
-    @Binding var outputSRTFilePath: String
-    @Binding var isPresented: Bool
-
-    var body: some View {
-        GeometryReader { geometry in
-            VStack() {
-                HStack{
-                    Text("Project: \(projectName)").font(.title2)
-                    Spacer()
-                    Button(action: {
-                        self.outputCaptions = ""
-                        self.progress = 0.0
-                        self.progressPercentage = 0
-                        self.remainingTime = "00:00"
-                        self.isPresented = false
-                        self.outputSRTFilePath = ""
-                        self.outputFCPXMLFilePath = ""
-                        self.totalBatch = 100000
-                        self.currentBatch = -100000
-                        self.status = "Spliting audio file by every 10 minutes···"
-                    }) {
-                        Image(systemName: "arrow.counterclockwise")
-                        Text("Reset")
-                    }
-                    
-                }.padding()
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.bottom, -28).padding(.top, -20)
-                    .disabled(self.currentBatch != self.totalBatch || self.progressPercentage < 100)
-                
-                Text("Current Status: \(self.status)").font(.title2)
-                    .padding()
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.bottom, -28)
-                
-                ScrollView {
-                    Text(outputCaptions)
-                        .font(.title3)
-                        .lineLimit(nil)
-                        .fixedSize(horizontal: false, vertical: true)
-                        .padding(.horizontal, -2)
-                        .padding(.vertical, -26)
-                        .padding()
-                        .frame(maxWidth: .infinity, alignment: .leading)
-                    
-                }.overlay(
-                    RoundedRectangle(cornerRadius: 10)
-                        .stroke(
-                            LinearGradient(
-                                gradient: Gradient(colors: [.orange, .red, .pink, .purple, .blue, .cyan, .green, .yellow]),
-                                startPoint: .leading,
-                                endPoint: .trailing
-                            ),
-                            lineWidth: 1
-                        )
-                )
-                .padding()
-                .padding(.bottom, -20)
-                .shadow(color: Color.gray.opacity(0.5), radius: 4, x: 0, y: 2)
-        
-                
-                HStack {
-                    Text("Download files: ").font(.title2)
-                    Button(action: {
-                        FileUtility.saveFileWithDialog(filePath: outputSRTFilePath)
-                    }) {
-                        Image(systemName: "square.and.arrow.down")
-                        Text("Download .srt file")
-                    }.buttonStyle(.borderedProminent).controlSize(.large)
-                        .tint(.purple)
-                        .disabled(self.currentBatch != self.totalBatch || self.progressPercentage < 100)
-
-                    Button(action: {
-                        FileUtility.saveFileWithDialog(filePath: outputFCPXMLFilePath)
-                    }) {
-                        Image(systemName: "square.and.arrow.down")
-                        Text("Download .fcpxml file")
-                    }.buttonStyle(.borderedProminent).controlSize(.large)
-                        .tint(.blue)
-                        .disabled(self.currentBatch != self.totalBatch || self.progressPercentage < 100)
-
-                    Button(action: {
-                        FileUtility.saveFileWithDialog(filePath: outputSRTFilePath)
-                        FileUtility.saveFileWithDialog(filePath: outputFCPXMLFilePath)
-                    }) {
-                        Image(systemName: "folder.badge.plus")
-                        Text("Download All")
-                    }.buttonStyle(.borderedProminent).controlSize(.large)
-                        .tint(.green)
-                        .disabled(self.currentBatch != self.totalBatch || self.progressPercentage < 100 || self.currentBatch == self.totalBatch - 1)
-                    
-                }
-                .padding().padding(.bottom, -20)
-                .frame(width: geometry.size.width,  alignment: .leading)
-                
-                
-                
-                HStack {
-                    Text("Open in Final Cut Pro: ")
-                        .font(.title2)
-                    Button(action: {
-//                        print("back to fcpx")
-                        backtofcpx(fcpxml_path: outputFCPXMLFilePath)
-                    }) {
-                        if let imagePath = Bundle.main.path(forResource: "fcpx-icon", ofType: "png"),
-                           let nsImage = NSImage(contentsOfFile: imagePath) {
-                            Image(nsImage: nsImage)
-                                .resizable()
-                                .aspectRatio(contentMode: .fit)
-                                .frame(width: nsImage.size.width * 0.05, height: nsImage.size.height * 0.05)
-                        }
-                        Text("Click here to check auto captions in Final Cut Pro X")
-                    }.buttonStyle(.borderedProminent)
-                        .controlSize(.large)
-                        .tint(.gray)
-                        .disabled(self.currentBatch != self.totalBatch || self.progressPercentage < 100)
-                }.padding().padding(.bottom, -20).frame(width: geometry.size.width,  alignment: .leading)
-                
-                LinearGradient(gradient: Gradient(colors: [.orange, .red, .pink, .purple, .blue, .cyan, .green, .yellow]), startPoint: .leading, endPoint: .trailing)
-                    .mask(
-                        GeometryReader { geometry in
-                            RoundedRectangle(cornerRadius: 6)
-                                .frame(width: geometry.size.width * CGFloat(progress), height: geometry.size.height)
-                        }
-                    )
-                
-                    .frame(height:8)
-                    .background(
-                        RoundedRectangle(cornerRadius: 6)
-                            .fill(
-                                LinearGradient(gradient: Gradient(colors: [.orange.opacity(0.3), .red.opacity(0.3), .pink.opacity(0.3), .purple.opacity(0.3), .blue.opacity(0.3), .cyan.opacity(0.3), .green.opacity(0.3), .yellow.opacity(0.3)]), startPoint: .leading, endPoint: .trailing)
-                            )
-                    ).padding().padding(.bottom, -16)
-                Text("Batch (\(self.currentBatch == -100000 ? "···" : String(self.currentBatch)) / \(totalBatch == 100000 ? "···" : String(totalBatch))): \(progressPercentage)% completed - \(remainingTime) remaining")
-    
-            }
-        }
-        .padding()
-
-    }
-
-    func backtofcpx(fcpxml_path: String) {
-        let command =
-        """
-        tell application "Final Cut Pro"
-            launch
-            activate
-            open POSIX file "\(fcpxml_path)"
-        end tell
-        """
-        DispatchQueue.global(qos: .background).async {
-            var error: NSDictionary?
-            if let scriptObject = NSAppleScript(source: command) {
-                _ = scriptObject.executeAndReturnError(&error)
-            }
-        }
-    }
-}
-
-// MARK: - SRT Converter Views
-struct SRTConverterView: View {
-    #if DEBUG
-    @ObserveInjection var inject
-    #endif
-    @State private var srtFileURL: URL?
-    @State private var fps: String = ""
-    @State private var selectedLanguage = "English"
-    @State private var projectName = ""
-    @State private var outputFCPXMLFilePath = ""
-    @State private var conversionComplete = false
-
-    let languages = ["Arabic", "Azerbaijani", "Armenian", "Albanian", "Afrikaans", "Amharic", "Assamese", "Bulgarian", "Bengali", "Breton", "Basque", "Bosnian", "Belarusian", "Bashkir", "Chinese Simplified", "Chinese Traditional", "Catalan", "Czech", "Croatian", "Dutch", "Danish", "English", "Estonian", "French", "Finnish", "Faroese", "German", "Greek", "Galician", "Georgian", "Gujarati", "Hindi", "Hebrew", "Hungarian", "Haitian creole", "Hawaiian", "Hausa", "Italian", "Indonesian", "Icelandic", "Japanese", "Javanese", "Korean", "Kannada", "Kazakh", "Khmer", "Lithuanian", "Latin", "Latvian", "Lao", "Luxembourgish", "Lingala", "Malay", "Maori", "Malayalam", "Macedonian", "Mongolian", "Marathi", "Maltese", "Myanmar", "Malagasy", "Norwegian", "Nepali", "Nynorsk", "Occitan", "Portuguese", "Polish", "Persian", "Punjabi", "Pashto", "Russian", "Romanian", "Spanish", "Swedish", "Slovak", "Serbian", "Slovenian", "Swahili", "Sinhala", "Shona", "Somali", "Sindhi", "Sanskrit", "Sundanese", "Turkish", "Tamil", "Thai", "Telugu", "Tajik", "Turkmen", "Tibetan", "Tagalog", "Tatar", "Ukrainian", "Urdu", "Uzbek", "Vietnamese", "Welsh", "Yoruba", "Yiddish"]
-
-    var body: some View {
-        Group {
-            if conversionComplete {
-                SRTConverterResultView(
-                    projectName: projectName,
-                    outputFCPXMLFilePath: $outputFCPXMLFilePath,
-                    conversionComplete: $conversionComplete,
-                    srtFileURL: $srtFileURL,
-                    fps: $fps,
-                    selectedLanguage: $selectedLanguage
-                )
-            } else {
-                SRTConverterInputView(
-                    srtFileURL: $srtFileURL,
-                    fps: $fps,
-                    selectedLanguage: $selectedLanguage,
-                    projectName: $projectName,
-                    outputFCPXMLFilePath: $outputFCPXMLFilePath,
-                    conversionComplete: $conversionComplete,
-                    languages: languages
-                )
-            }
-        }
-
-    }
-}
-
-struct SRTConverterInputView: View {
-    #if DEBUG
-    @ObserveInjection var inject
-    #endif
-    @Binding var srtFileURL: URL?
-    @Binding var fps: String
-    @Binding var selectedLanguage: String
-    @Binding var projectName: String
-    @Binding var outputFCPXMLFilePath: String
-    @Binding var conversionComplete: Bool
-    let languages: [String]
-
-    @State private var fileName: String = ""
-    @State private var selectedResolution: VideoResolution = .fullHD1080p
-    @State private var customWidth: String = "1920"
-    @State private var customHeight: String = "1080"
-    @State private var selectedFrameRate: FrameRate = .fps30
-    @State private var customFps: String = "30"
-
-    private var currentWidth: Int {
-        if selectedResolution == .custom {
-            return Int(customWidth) ?? 1920
-        }
-        return selectedResolution.width
-    }
-
-    private var currentHeight: Int {
-        if selectedResolution == .custom {
-            return Int(customHeight) ?? 1080
-        }
-        return selectedResolution.height
-    }
-
-    private var currentFps: Float {
-        if selectedFrameRate == .custom {
-            return Float(customFps) ?? 30.0
-        }
-        return selectedFrameRate.value
-    }
-
-    private var isResolutionValid: Bool {
-        let validation = VideoResolution.isValidResolution(width: currentWidth, height: currentHeight)
-        return validation.valid
-    }
-
-    private var resolutionWarning: String? {
-        let validation = VideoResolution.isValidResolution(width: currentWidth, height: currentHeight)
-        return validation.message
-    }
-
-    private var isFpsValid: Bool {
-        return FrameRate.isValidFrameRate(currentFps)
-    }
-
-    var body: some View {
-        VStack {
-            Grid(alignment: .leadingFirstTextBaseline, verticalSpacing: 20) {
-                GridRow {
-                    Text(String(localized: "SRT File:", comment: "SRT file label"))
-                        .gridColumnAlignment(.trailing)
-                    HStack {
-                        Button(action: {
-                            let panel = NSOpenPanel()
-                            panel.canChooseFiles = true
-                            panel.canChooseDirectories = false
-                            panel.allowsMultipleSelection = false
-                            panel.allowedContentTypes = [.init(filenameExtension: "srt")!]
-                            if panel.runModal() == .OK {
-                                self.srtFileURL = panel.urls.first
-                                if let url = panel.urls.first {
-                                    self.fileName = url.lastPathComponent
-                                    self.projectName = url.deletingPathExtension().lastPathComponent
-                                }
-                            }
-                        }) {
-                            Text(String(localized: "Choose File", comment: "Choose file button"))
-                        }
-
-                        if srtFileURL != nil {
-                            Text(fileName)
-                        }
-                    }
-                    .gridColumnAlignment(.leading)
-                }
-
-                GridRow {
-                    Text(String(localized: "Project Name:", comment: "Project name label"))
-                    TextField(String(localized: "Project name for FCPXML", comment: "Project name placeholder"), text: $projectName)
-                        .textFieldStyle(RoundedBorderTextFieldStyle())
-                        .frame(width: 200)
-                }
-
-                GridRow {
-                    Text(String(localized: "Frame Rate:", comment: "Frame rate label"))
-                    VStack(alignment: .leading, spacing: 8) {
-                        Picker(selection: $selectedFrameRate, label: EmptyView()) {
-                            ForEach(FrameRate.allCases) { frameRate in
-                                Text(frameRate.displayName).tag(frameRate)
-                            }
-                        }
-                        .pickerStyle(MenuPickerStyle())
-                        .frame(alignment: .leading)
-
-                        if selectedFrameRate == .custom {
-                            HStack(spacing: 8) {
-                                TextField("fps", text: $customFps)
-                                    .textFieldStyle(RoundedBorderTextFieldStyle())
-                                    .frame(width: 70)
-                                Text("fps")
-                            }
-
-                            if !isFpsValid {
-                                Text(String(localized: "Frame rate must be between 0 and 120", comment: "Invalid frame rate warning"))
-                                    .font(.caption)
-                                    .foregroundColor(.red)
-                            }
-                        }
-                    }
-                }
-
-                GridRow {
-                    Text(String(localized: "Resolution:", comment: "Resolution label"))
-                    VStack(alignment: .leading, spacing: 8) {
-                        Picker(selection: $selectedResolution, label: EmptyView()) {
-                            ForEach(VideoResolution.allCases) { resolution in
-                                Text(resolution.displayName).tag(resolution)
-                            }
-                        }
-                        .pickerStyle(MenuPickerStyle())
-                        .frame(alignment: .leading)
-
-                        if selectedResolution == .custom {
-                            HStack(spacing: 8) {
-                                TextField(String(localized: "Width", comment: "Width placeholder"), text: $customWidth)
-                                    .textFieldStyle(RoundedBorderTextFieldStyle())
-                                    .frame(width: 70)
-                                Text("×")
-                                TextField(String(localized: "Height", comment: "Height placeholder"), text: $customHeight)
-                                    .textFieldStyle(RoundedBorderTextFieldStyle())
-                                    .frame(width: 70)
-                            }
-
-                            if let warning = resolutionWarning {
-                                Text(warning)
-                                    .font(.caption)
-                                    .foregroundColor(isResolutionValid ? .orange : .red)
-                            }
-                        }
-                    }
-                }
-
-                GridRow {
-                    Text(String(localized: "Language:", comment: "Language label"))
-                    Picker(selection: $selectedLanguage, label: EmptyView()) {
-                        ForEach(languages, id: \.self) {
-                            Text($0)
-                        }
-                    }
-                    .pickerStyle(MenuPickerStyle())
-                    .frame(alignment: .leading)
-                }
-
-                GridRow {
-                    Button(action: {
-                        convertSRTtoFCPXML()
-                    }, label: {
-                        Text(String(localized: "Convert to FCPXML", comment: "Convert button label"))
-                    })
-                    .buttonStyle(BorderedProminentButtonStyle())
-                    .gridCellAnchor(.center)
-                    .disabled(srtFileURL == nil || projectName.isEmpty || !isResolutionValid || !isFpsValid)
-                }.gridCellColumns(2)
-            }
-            .padding()
-        }
-
-    }
-
-    private func convertSRTtoFCPXML() {
-        guard let srtURL = srtFileURL else { return }
-
-        let srtPath = srtURL.path
-        outputFCPXMLFilePath = SRTConverter.srtToFCPXML(
-            srtPath: srtPath,
-            fps: currentFps,
-            projectName: projectName,
-            language: selectedLanguage,
-            width: currentWidth,
-            height: currentHeight
-        )
-
-        if outputFCPXMLFilePath != "Error" {
-            conversionComplete = true
-        }
-    }
-}
-
-struct SRTConverterResultView: View {
-    #if DEBUG
-    @ObserveInjection var inject
-    #endif
-    let projectName: String
-    @Binding var outputFCPXMLFilePath: String
-    @Binding var conversionComplete: Bool
-    @Binding var srtFileURL: URL?
-    @Binding var fps: String
-    @Binding var selectedLanguage: String
-
-    var body: some View {
-        VStack(spacing: 20) {
-            Image(systemName: "checkmark.circle.fill")
-                .resizable()
-                .frame(width: 60, height: 60)
-                .foregroundColor(.green)
-
-            Text("Conversion Complete!")
-                .font(.title)
-
-            Text("Project: \(projectName)")
-                .font(.title2)
-
-            HStack(spacing: 16) {
-                Button(action: {
-                    FileUtility.saveFileWithDialog(filePath: outputFCPXMLFilePath)
-                }) {
-                    Image(systemName: "square.and.arrow.down")
-                    Text("Download .fcpxml")
-                }
-                .buttonStyle(.borderedProminent)
-                .controlSize(.large)
-                .tint(.blue)
-
-                Button(action: {
-                    backtofcpx(fcpxml_path: outputFCPXMLFilePath)
-                }) {
-                    if let imagePath = Bundle.main.path(forResource: "fcpx-icon", ofType: "png"),
-                       let nsImage = NSImage(contentsOfFile: imagePath) {
-                        Image(nsImage: nsImage)
-                            .resizable()
-                            .aspectRatio(contentMode: .fit)
-                            .frame(width: nsImage.size.width * 0.05, height: nsImage.size.height * 0.05)
-                    }
-                    Text("Open in Final Cut Pro")
-                }
-                .buttonStyle(.borderedProminent)
-                .controlSize(.large)
-                .tint(.gray)
-            }
-
-            Button(action: {
-                resetState()
-            }) {
-                Image(systemName: "arrow.counterclockwise")
-                Text("Convert Another File")
-            }
-            .buttonStyle(.bordered)
-            .controlSize(.large)
-        }
-        .padding()
-
-    }
-
-    private func resetState() {
-        conversionComplete = false
-        outputFCPXMLFilePath = ""
-        srtFileURL = nil
-        fps = ""
-        selectedLanguage = "English"
-    }
-
-    private func backtofcpx(fcpxml_path: String) {
-        let command =
-        """
-        tell application "Final Cut Pro"
-            launch
-            activate
-            open POSIX file "\(fcpxml_path)"
-        end tell
-        """
-        DispatchQueue.global(qos: .background).async {
-            var error: NSDictionary?
-            if let scriptObject = NSAppleScript(source: command) {
-                _ = scriptObject.executeAndReturnError(&error)
-            }
-        }
-    }
-}
-
+// MARK: - Preview
 struct ContentView_Previews: PreviewProvider {
     static var previews: some View {
         ContentView()
     }
 }
-
-
