+++ Whisper Auto Captions/Views/Transcription/HomeView.swift
@@ -1,0 +1,163 @@
+import SwiftUI
+#if DEBUG
+import Inject
+#endif
+
+// MARK: - Home View
+struct HomeView: View {
+    #if DEBUG
+    @ObserveInjection var inject
+    #endif
+    
+    @ObservedObject var viewModel: HomeViewModel
+    @ObservedObject private var settingsManager = SettingsManager.shared
+
+    var body: some View {
+        VStack {
+            Grid(alignment: .leadingFirstTextBaseline, verticalSpacing: 20) {
+                // Audio File Selection
+                GridRow {
+                    Text("Audio File:")
+                        .gridColumnAlignment(.trailing)
+                    HStack {
+                        Button(action: selectFile) {
+                            Text("Choose File")
+                        }
+
+                        if viewModel.fileURL != nil {
+                            Text(viewModel.fileName)
+                        }
+                    }
+                    .gridColumnAlignment(.leading)
+                }
+
+                // Frame Rate Selection
+                GridRow {
+                    Text(String(localized: "Frame Rate:", comment: "Frame rate label"))
+                    VStack(alignment: .leading, spacing: 8) {
+                        Picker(selection: $viewModel.selectedFrameRate, label: EmptyView()) {
+                            ForEach(FrameRate.allCases) { frameRate in
+                                Text(frameRate.displayName).tag(frameRate)
+                            }
+                        }
+                        .pickerStyle(MenuPickerStyle())
+                        .frame(alignment: .leading)
+
+                        if viewModel.selectedFrameRate == .custom {
+                            HStack(spacing: 8) {
+                                TextField("fps", text: $viewModel.customFps)
+                                    .textFieldStyle(RoundedBorderTextFieldStyle())
+                                    .frame(width: 70)
+                                Text("fps")
+                            }
+
+                            if !viewModel.isFpsValid {
+                                Text(String(localized: "Frame rate must be between 0 and 120", comment: "Invalid frame rate warning"))
+                                    .font(.caption)
+                                    .foregroundColor(.red)
+                            }
+                        }
+                    }
+                }
+
+                // Model Selection
+                GridRow {
+                    Text(String(localized: "Model:", comment: "Model label"))
+                    Picker(selection: $viewModel.selectedModel, label: EmptyView()) {
+                        ForEach(viewModel.models, id: \.self) {
+                            Text($0)
+                        }
+                    }
+                    .pickerStyle(MenuPickerStyle())
+                    .frame(alignment: .leading)
+                }
+
+                // Language Selection
+                GridRow {
+                    Text(String(localized: "Language:", comment: "Language label"))
+                    Picker(selection: $viewModel.selectedLanguage, label: EmptyView()) {
+                        ForEach(viewModel.languages, id: \.self) {
+                            Text($0)
+                        }
+                    }
+                    .pickerStyle(MenuPickerStyle())
+                    .frame(alignment: .leading)
+                }
+
+                // Preset Selection
+                GridRow {
+                    Text(String(localized: "Preset:", comment: "Preset label"))
+                    Picker(selection: $viewModel.selectedPreset, label: EmptyView()) {
+                        ForEach(WhisperPreset.allCases) { preset in
+                            Text(preset.displayName).tag(preset)
+                        }
+                    }
+                    .pickerStyle(MenuPickerStyle())
+                    .frame(alignment: .leading)
+                    .onChange(of: viewModel.selectedPreset) { newPreset in
+                        settingsManager.currentPreset = newPreset
+                    }
+                }
+
+                // Action Buttons
+                GridRow {
+                    HStack(spacing: 12) {
+                        Button(action: { viewModel.showSettings = true }) {
+                            HStack {
+                                Image(systemName: "gear")
+                                Text(String(localized: "Settings", comment: "Settings button"))
+                            }
+                        }
+
+                        Button(action: {
+                            viewModel.validateAndStartTranscription()
+                        }, label: {
+                            Text(String(localized: "Create", comment: "Create button"))
+                        })
+                        .buttonStyle(BorderedProminentButtonStyle())
+                        .disabled(viewModel.fileURL == nil || !viewModel.isFpsValid)
+                    }
+                }
+                .gridCellColumns(2)
+                .gridCellAnchor(.center)
+            }
+            
+            if viewModel.isDownloading {
+                ProgressView(value: viewModel.downloadProgress)
+                    .padding()
+                    .progressViewStyle(LinearProgressViewStyle())
+                    .frame(width: 200)
+            }
+        }
+        .alert(isPresented: $viewModel.showAlert) {
+            Alert(
+                title: Text("Downloading \(viewModel.selectedModel) Model"),
+                message: Text(String(format: "Progress: %.0f%%", viewModel.downloadProgress * 100)),
+                primaryButton: .destructive(Text("Cancel"), action: {
+                    viewModel.cancelDownload()
+                }),
+                secondaryButton: .default(Text(""), action: {})
+            )
+        }
+        .sheet(isPresented: $viewModel.showSettings) {
+            SettingsWindowView()
+        }
+        #if DEBUG
+        .enableInjection()
+        #endif
+    }
+    
+    // MARK: - Actions
+    private func selectFile() {
+        let panel = NSOpenPanel()
+        panel.canChooseFiles = true
+        panel.canChooseDirectories = true
+        panel.allowsMultipleSelection = false
+        panel.allowedContentTypes = [.audio, .movie]
+        if panel.runModal() == .OK {
+            if let url = panel.urls.first {
+                viewModel.selectFile(url: url)
+            }
+        }
+    }
+}
